[hip] Move approx_atan2.h to CUDADataFormats because std::round is not constexpr in HIP
