+++ src/hip/CUDADataFormats/approx_atan2.h
